Fix an issue that all words are deleted when there is no valid input data
Fix an issue that dict words in the file pw_dict are deleted when
create-cracklib-dict or cracklib-packer has no valid input data.

For example:
Give an empty file or a file that does't exist or a file has no valid
input words to create-cracklib-dict, all words in pw_dict are deleted.
$ cracklib-create-dict test

Run cracklib-packer and press Enter, all words in pw_dict are deleted.
$ cracklib-packer