Remove initialization of buffer area
The code that introduced this line contained an important bug fix,
doubling the size of the working buffer, which was the correct fix.
However it also initialized the memory which is neither necessary nor
useful and considerably reduces performance which is easily and safely
restored by not initializing the buffer.

Signed-off-by: Ed Beroset <beroset@ieee.org>